requirements.txt
face_recognition
Pillow